+++ Services/Verient/IModelRepository.cs
@@ -6,7 +6,7 @@ namespace VCA.Services.Verient
 {
     public interface IModelRepository
     {
-        Task<List<Model>> FindByManufacturerIdAndSegmentIdAsync(long manuId, long segId, int page = 1, int pageSize = 10);
+        Task<List<Model>> FindByManufacturerIdAndSegmentIdAsync(int manuId, int segId, int page = 1, int pageSize = 10);
 
         Task<List<Model>> FindAllByOrderByCreatedAtDescAsync(int page = 1, int pageSize = 10);
     }
